grocegoo - Web application
Groceroo is a web application that socializes the grocery shopping
experience and it is our answer to the problem. It features the ability
to let users share their grocery lists with others in a location based
communal grocery shopping network and makes acquiring groceries more
time-efficient for everyone in the network.

Credit:
ADIB KASHEM
CONNIR LEMOS
HONG LE
TRIMAN NGUYEN
YANG ZHAO